ADTsbase replaced with ADTsCore - Only contains neccessary ADT features for RelationTerm so far. RelationTerm now compiling.
