+++ Project.toml
@@ -11,6 +11,7 @@ Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 CompTime = "0fb5dd42-039a-4ca4-a1d7-89a96eae6d39"
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
+DiagrammaticEquations = "6f00c28b-6bed-4403-80fa-30e0dc12f317"
 GATlab = "f0ffcf3b-d13a-433e-917c-cc44ccf5ead2"
 GeneralizedGenerated = "6b9d7cbe-bcb9-11e9-073f-15a7a543e2eb"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
